Pull the messy promise wrapping into an internal wrapper file
Someone making changes to `Database` shouldn't need to wrangle this code
unless they really need to. Moving it into a wrapper makes `Database.ts`
a bit friendlier to look at.
